refactor: remove to move shared pagination
