+++ src/feature/post/ui/PostList.tsx
@@ -2,7 +2,7 @@ import { useQuery, useQueryClient } from "@tanstack/react-query"
 import React, { useEffect, useState } from "react"
 import { getTags, getUser, Tags } from "../../../entities"
 import { userPostInfo, useSearchQueryStore, useSelectedUserStore } from "../model"
-import { Pagination } from "../../../widgets"
+import { Pagination } from "../../../shared"
 import { Author, CardContent, QUERY_KEYS, useURL } from "../../../shared"
 import { PostSearchFilter } from "./PostSearchFilter"
 import { PostTable } from "./PostTable"
