+++ src/shared/ui/index.ts
@@ -22,3 +22,4 @@ export { DialogContent } from "./dialog/dialog-content"
 export { DialogHeader } from "./dialog/dialog-header"
 export { DialogTitle } from "./dialog/dialog-title"
 export { HighlightText } from "./hight-text"
+export {Pagination} from './pagination'
