+++ src/shared/ui/pagination/type.ts
