+++ src/widgets/index.ts
@@ -1,3 +1,2 @@
 export { Header } from "./Header"
 export { Footer } from "./Footer"
-export { Pagination } from "./pagination"
